Moved CurrentLocationButton component into LocationSearch folder
